+++ apps/laboratory/src/pages/index.tsx
@@ -11,7 +11,12 @@ import {
   Link
 } from '@chakra-ui/react'
 import { IoArrowForward } from 'react-icons/io5'
-import { wagmiSdkOptions, ethersSdkOptions, solanaSdkOptions } from '../utils/DataUtil'
+import {
+  wagmiSdkOptions,
+  ethersSdkOptions,
+  solanaSdkOptions,
+  ethers5SdkOptions
+} from '../utils/DataUtil'
 import { RandomLink } from '../components/RandomLink'
 
 export default function HomePage() {
@@ -99,6 +104,34 @@ export default function HomePage() {
         <CardBody>
           <Stack divider={<StackDivider />} spacing="4">
             {ethersSdkOptions.map(option => (
+              <Box key={option.link}>
+                <Stack direction="row" justifyContent="space-between" alignItems="center">
+                  <Box>
+                    <Heading size="xs" textTransform="uppercase">
+                      {option.title}
+                    </Heading>
+                    <Text pt="2" fontSize="sm">
+                      {option.description}
+                    </Text>
+                  </Box>
+                  <Link href={option.link}>
+                    <Button rightIcon={<IoArrowForward />}>Go</Button>
+                  </Link>
+                </Stack>
+              </Box>
+            ))}
+          </Stack>
+        </CardBody>
+      </Card>
+
+      <Card marginTop={10} marginBottom={10}>
+        <CardHeader>
+          <Heading size="md">Ethers5</Heading>
+        </CardHeader>
+
+        <CardBody>
+          <Stack divider={<StackDivider />} spacing="4">
+            {ethers5SdkOptions.map(option => (
               <Box key={option.link}>
                 <Stack direction="row" justifyContent="space-between" alignItems="center">
                   <Box>