+++ apps/laboratory/src/pages/library/ethers5-all.tsx
@@ -1,0 +1,40 @@
+import { createWeb3Modal, defaultConfig } from '@web3modal/ethers5/react'
+import { ThemeStore } from '../../utils/StoreUtil'
+import { EthersConstants } from '../../utils/EthersConstants'
+import { ConstantsUtil } from '../../utils/ConstantsUtil'
+import { AppKitButtons } from '../../components/AppKitButtons'
+import { siweConfig } from '../../utils/SiweUtils'
+import { SiweData } from '../../components/Siwe/SiweData'
+import { Ethers5Tests } from '../../components/Ethers/Ethers5Tests'
+import { Ethers5ModalInfo } from '../../components/Ethers/Ethers5ModalInfo'
+
+const modal = createWeb3Modal({
+  ethersConfig: defaultConfig({
+    metadata: ConstantsUtil.Metadata,
+    defaultChainId: 1,
+    auth: {
+      socials: ['google', 'x', 'discord', 'farcaster', 'github', 'apple', 'facebook']
+    },
+    coinbasePreference: 'smartWalletOnly'
+  }),
+  chains: EthersConstants.chains,
+  projectId: ConstantsUtil.ProjectId,
+  enableAnalytics: true,
+  metadata: ConstantsUtil.Metadata,
+  termsConditionsUrl: 'https://walletconnect.com/terms',
+  privacyPolicyUrl: 'https://walletconnect.com/privacy',
+  siweConfig
+})
+
+ThemeStore.setModal(modal)
+
+export default function Ethers() {
+  return (
+    <>
+      <AppKitButtons />
+      <Ethers5ModalInfo />
+      <SiweData />
+      <Ethers5Tests />
+    </>
+  )
+}