+++ apps/laboratory/tests/siwe-email.spec.ts
@@ -65,14 +65,15 @@ emailSiweTest('it should reject sign', async () => {
   await validator.expectRejectedSign()
 })
 
-emailSiweTest('it should switch network and sign', async () => {
+emailSiweTest('it should switch network and sign', async ({ library }) => {
   let targetChain = 'Polygon'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.switchNetwork(targetChain)
   await page.promptSiwe()
   await page.approveSign()
+  if (library === 'wagmi') {
+    await page.goToSettings()
+  }
   await validator.expectSwitchedNetwork(targetChain)
   await page.closeModal()
   await page.sign()
@@ -80,12 +81,13 @@ emailSiweTest('it should switch network and sign', async () => {
   await validator.expectAcceptedSign()
 
   targetChain = 'Ethereum'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.switchNetwork(targetChain)
   await page.promptSiwe()
   await page.approveSign()
+  if (library === 'wagmi') {
+    await page.goToSettings()
+  }
   await validator.expectSwitchedNetwork(targetChain)
   await page.closeModal()
   await page.sign()