+++ packages/core/src/controllers/RouterController.ts
@@ -2,6 +2,7 @@ import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 import { proxy } from 'valtio/vanilla'
 import type { CaipNetwork, Connector, WcWallet } from '../utils/TypeUtil.js'
 import type { SwapInputTarget } from './SwapController.js'
+import type { Chain } from '@web3modal/common'
 
 // -- Types --------------------------------------------- //
 type TransactionAction = {
@@ -61,6 +62,8 @@ export interface RouterControllerState {
     | 'Swap'
     | 'SwapSelectToken'
     | 'SwapPreview'
+    | 'ConnectingMultiChain'
+    | 'SwitchActiveChain'
   history: RouterControllerState['view'][]
   data?: {
     connector?: Connector
@@ -70,6 +73,10 @@ export interface RouterControllerState {
     newEmail?: string
     target?: SwapInputTarget
     swapUnsupportedChain?: boolean
+    connectors?: Connector[]
+    switchToChain?: Chain
+    navigateTo?: RouterControllerState['view']
+    navigateWithReplace?: boolean
   }
   transactionStack: TransactionAction[]
 }