+++ packages/scaffold-ui/src/views/w3m-unsupported-chain-view/index.ts
@@ -140,7 +140,7 @@ export class W3mUnsupportedChainView extends LitElement {
         RouterController.push('SwitchNetwork', { ...routerData, network })
       }
     } else if (!isConnected) {
-      NetworkController.setCaipNetwork(network)
+      NetworkController.setActiveCaipNetwork(network)
       RouterController.push('Connect')
     }
   }