+++ packages/solana/exports/react.tsx
@@ -3,44 +3,55 @@
 import { useSnapshot } from 'valtio'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import { getWeb3Modal } from '@web3modal/scaffold-react'
+import { AppKit } from '@web3modal/base'
+import { SolanaWeb3JsClient, SolStoreUtil } from '@web3modal/base/adapters/solana/web3js'
+import { type Connection, type Provider } from '@web3modal/base/adapters/solana/web3js'
+import type { SolanaAppKitOptions } from './options'
 
-import { Web3Modal } from '../src/client.js'
+// -- Configs -----------------------------------------------------------
+export { defaultSolanaConfig } from '@web3modal/base/adapters/solana/web3js'
 
-import { SolStoreUtil, type Connection, type Provider } from '../src/utils/scaffold/index.js'
+// -- Setup -------------------------------------------------------------
+let appkit: AppKit | undefined = undefined
+let solanaAdapter: SolanaWeb3JsClient | undefined = undefined
 
-import type { Web3ModalOptions } from '../src/client.js'
+export function createWeb3Modal(options: SolanaAppKitOptions) {
+  solanaAdapter = new SolanaWeb3JsClient({
+    solanaConfig: options.solanaConfig,
+    chains: options.chains,
+    wallets: options.wallets,
+    projectId: options.projectId
+  })
+  appkit = new AppKit({
+    ...options,
+    adapters: [solanaAdapter],
+    sdkType: 'w3m',
+    sdkVersion: `react-solana-${ConstantsUtil.VERSION}`
+  })
+  getWeb3Modal(appkit)
 
-// -- Setup -------------------------------------------------------------------
-let modal: Web3Modal | undefined = undefined
-
-// -- Types -------------------------------------------------------------------
-export type { Web3Modal, Web3ModalOptions } from '../src/client.js'
-
-export function createWeb3Modal(options: Web3ModalOptions) {
-  if (!modal) {
-    modal = new Web3Modal({
-      ...options,
-      _sdkVersion: `react-solana-${ConstantsUtil.VERSION}`
-    })
-  }
-  getWeb3Modal(modal)
-
-  return modal
+  return appkit
 }
 
 // -- Hooks -------------------------------------------------------------------
-export function useWeb3ModalProvider() {
-  const { provider, connection } = useSnapshot(SolStoreUtil.state)
+export function useWeb3ModalProvider(): {
+  walletProvider: Provider | undefined
+  connection: Connection | undefined
+} {
+  const state = useSnapshot(SolStoreUtil.state)
 
   return {
-    walletProvider: provider as Provider | undefined,
-    connection: connection as Connection | undefined
+    walletProvider: state.provider,
+    connection: state.connection
+  } as {
+    walletProvider: Provider | undefined
+    connection: Connection | undefined
   }
 }
 
 export function useDisconnect() {
   function disconnect() {
-    modal?.disconnect()
+    solanaAdapter?.disconnect()
   }
 
   return {
@@ -65,6 +76,3 @@ export {
   useWeb3ModalState,
   useWeb3ModalEvents
 } from '@web3modal/scaffold-react'
-
-// -- Universal Exports -------------------------------------------------------
-export { defaultSolanaConfig } from '../src/utils/defaultConfig.js'