+++ packages/solana/exports/vue.ts
@@ -1,41 +1,44 @@
 import { onUnmounted, ref } from 'vue'
-
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import { getWeb3Modal } from '@web3modal/scaffold-vue'
-
-import type { Web3ModalOptions } from '../src/client.js'
-import type { CaipNetwork } from '@web3modal/scaffold'
-import { SolStoreUtil } from '../src/utils/scaffold/SolanaStoreUtil.js'
-import { Web3Modal } from '../src/client.js'
-
-// -- Types -------------------------------------------------------------------
-export type { Web3ModalOptions } from '../src/client.js'
+import { AppKit } from '@web3modal/base'
+import { SolanaWeb3JsClient, SolStoreUtil } from '@web3modal/base/adapters/solana/web3js'
+import type { Provider, Connection } from '@web3modal/base/adapters/solana/web3js'
+import type { CaipNetwork } from '@web3modal/core'
+import type { SolanaAppKitOptions } from './options'
 
 // -- Setup -------------------------------------------------------------------
-let modal: Web3Modal | undefined = undefined
+let appkit: AppKit | undefined = undefined
+let solanaAdapter: SolanaWeb3JsClient | undefined = undefined
 
-export function createWeb3Modal(options: Web3ModalOptions) {
-  if (!modal) {
-    modal = new Web3Modal({
-      ...options,
-      _sdkVersion: `vue-solana-${ConstantsUtil.VERSION}`
-    })
-    getWeb3Modal(modal)
-  }
+export function createWeb3Modal(options: SolanaAppKitOptions) {
+  solanaAdapter = new SolanaWeb3JsClient({
+    solanaConfig: options.solanaConfig,
+    chains: options.chains,
+    wallets: options.wallets,
+    projectId: options.projectId
+  })
+  appkit = new AppKit({
+    ...options,
+    adapters: [solanaAdapter],
+    sdkType: 'w3m',
+    sdkVersion: `vue-solana-${ConstantsUtil.VERSION}`
+  })
+  getWeb3Modal(appkit)
 
-  return modal
+  return appkit
 }
 
 // -- Composites --------------------------------------------------------------
 export function useWeb3ModalProvider() {
-  if (!modal) {
+  if (!solanaAdapter) {
     throw new Error('Please call "createWeb3Modal" before using "useWeb3ModalProvider" composition')
   }
 
   const walletProvider = ref(SolStoreUtil.state.provider)
   const connection = ref(SolStoreUtil.state.connection)
 
-  const unsubscribe = modal.subscribeProvider(state => {
+  const unsubscribe = solanaAdapter.subscribeProvider(state => {
     walletProvider.value = state.provider
   })
 
@@ -46,12 +49,15 @@ export function useWeb3ModalProvider() {
   return {
     walletProvider: walletProvider.value ?? undefined,
     connection: connection.value ?? undefined
+  } as {
+    walletProvider: Provider | undefined
+    connection: Connection | undefined
   }
 }
 
 export function useDisconnect() {
   function disconnect() {
-    modal?.disconnect()
+    solanaAdapter?.disconnect()
   }
 
   return {
@@ -61,7 +67,7 @@ export function useDisconnect() {
 
 export function useSwitchNetwork() {
   async function switchNetwork(chainId: string) {
-    await modal?.switchNetwork({ id: chainId } as CaipNetwork)
+    await solanaAdapter?.switchNetwork({ id: chainId } as CaipNetwork)
   }
 
   return {
@@ -70,15 +76,15 @@ export function useSwitchNetwork() {
 }
 
 export function useWeb3ModalAccount() {
-  if (!modal) {
+  if (!solanaAdapter) {
     throw new Error('Please call "createWeb3Modal" before using "useWeb3ModalAccount" composition')
   }
 
-  const address = ref(modal.getAddress())
+  const address = ref(solanaAdapter.getAddress())
   const isConnected = ref(SolStoreUtil.state.isConnected)
   const chainId = ref(SolStoreUtil.state.currentChain?.chainId)
 
-  const unsubscribe = modal.subscribeProvider(state => {
+  const unsubscribe = solanaAdapter.subscribeProvider(state => {
     address.value = state.address ?? ''
     isConnected.value = state.isConnected
     chainId.value = state.chainId
@@ -96,13 +102,13 @@ export function useWeb3ModalAccount() {
 }
 
 export function useWeb3ModalError() {
-  if (!modal) {
+  if (!solanaAdapter) {
     throw new Error('Please call "createWeb3Modal" before using "useWeb3ModalError" composition')
   }
 
   const error = ref(SolStoreUtil.state.error)
 
-  const unsubscribe = modal.subscribeProvider(state => {
+  const unsubscribe = solanaAdapter.subscribeProvider(state => {
     error.value = state.error
   })
 
@@ -123,4 +129,4 @@ export {
 } from '@web3modal/scaffold-vue'
 
 // -- Universal Exports -------------------------------------------------------
-export { defaultSolanaConfig } from '../src/utils/defaultConfig.js'
+export { defaultSolanaConfig } from '@web3modal/base/adapters/solana/web3js'