+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @web3modal/ethers
 
+## 5.1.1
+
+### Patch Changes
+
+- Update EthProvider to v 2.15.1
+
+- Updated dependencies []:
+  - @web3modal/common@5.1.1
+  - @web3modal/polyfills@5.1.1
+  - @web3modal/scaffold@5.1.1
+  - @web3modal/scaffold-react@5.1.1
+  - @web3modal/scaffold-utils@5.1.1
+  - @web3modal/scaffold-vue@5.1.1
+  - @web3modal/siwe@5.1.1
+  - @web3modal/wallet@5.1.1
+
 ## 5.1.0
 
 ### Minor Changes